src/test/kotlin/mpdev/springboot/aoc2017/day22/Day22Test.kt
package mpdev.springboot.aoc2017.day22

import mpdev.springboot.aoc2017.input.InputDataReader
import mpdev.springboot.aoc2017.solutions.day22.Day22
import mpdev.springboot.aoc2017.solutions.day22.Virus
import mpdev.springboot.aoc2017.utils.Point
import org.assertj.core.api.Assertions.assertThat
import org.junit.jupiter.api.BeforeEach
import org.junit.jupiter.api.Order
import org.junit.jupiter.api.Test
import org.junit.jupiter.api.TestInstance

@TestInstance(TestInstance.Lifecycle.PER_CLASS)
class Day22Test {

    private val day = 22                                     ///////// Update this for a new dayN test
    private val puzzleSolver = Day22()                      ///////// Update this for a new dayN test
    private val inputDataReader = InputDataReader("src/test/resources/inputdata/input")
    private var inputLines: List<String> = inputDataReader.read(day)

    @BeforeEach
    fun setup() {
        puzzleSolver.setDay()
        puzzleSolver.inputData = inputLines
        puzzleSolver.initSolver()
    }

    @Test
    @Order(1)
    fun `Sets Day correctly`() {
        assertThat(puzzleSolver.day).isEqualTo(day)
    }

    @Test
    @Order(2)
    fun `Reads Input and sets up Grid`() {
        val virus = Virus(inputLines)
        virus.grid.print()
        assertThat(virus.start).isEqualTo(Point(1,1))
    }

    @Test
    @Order(3)
    fun `Simulates Particle Movement`() {
        val virus = Virus(inputLines)
        val result = virus.solve1(70)
        virus.grid.updateDimensions()
        virus.grid.print()
        assertThat(result).isEqualTo(41)
    }

    @Test
    @Order(5)
    fun `Solves Part 1`() {
        assertThat(puzzleSolver.solvePart1().result).isEqualTo("5587")
    }

    @Test
    @Order(7)
    fun `Solves Part 2`() {
        assertThat(puzzleSolver.solvePart2().result).isEqualTo("2511944")
    }

}
